pong/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-environ==0.11.2
django-cors-headers
environ==1.0
idna==3.7
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
psycopg2-binary==2.9.9
djangorestframework
drf-spectacular
python-decouple==3.8
python-environ
channels
daphne
channels_redis
pillow
web3[tester]
python-dotenv
eth-utils
py-evm
selenium
